moved repetable logic to utils
