--- feature/client/src/desktopMain/kotlin/com/mifos/feature/client/clientDetails/ClientDetailsScreen.desktop.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/android-client/blob/master/LICENSE.md
- */
-package com.mifos.feature.client.clientDetails
-
-import androidx.compose.runtime.Composable
-import com.mifos.feature.client.utils.PlatformCameraLauncher
-
-@Composable
-actual fun rememberPlatformCameraLauncher(onImageCapturedPath: (String?) -> Unit): PlatformCameraLauncher {
-    TODO("Not yet implemented")
-}
